test: Create utility and update tests
